@@ -3,4 +3,4 @@ version: 0.0.1
 description: An example of using DynamicValue.
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
